wip(deps): add deps component, collecting brick imports
