+++ components/polylith/deps/__init__.py
@@ -1,0 +1,3 @@
+from polylith.deps.core import get_brick_imports
+
+__all__ = ["get_brick_imports"]
