+++ components/polylith/sync/collect.py
@@ -1,26 +1,6 @@
 from pathlib import Path
 
-from polylith import check, info, workspace
-
-
-def get_brick_imports(root: Path, ns: str, project_data: dict) -> dict:
-    bases = {b for b in project_data.get("bases", [])}
-    components = {c for c in project_data.get("components", [])}
-
-    bases_paths = workspace.paths.collect_bases_paths(root, ns, bases)
-    components_paths = workspace.paths.collect_components_paths(root, ns, components)
-
-    brick_imports_in_bases = check.collect.extract_bricks(bases_paths, ns)
-    brick_imports_in_components = check.collect.extract_bricks(components_paths, ns)
-
-    return {
-        "bases": check.collect.with_unknown_components(
-            root, ns, brick_imports_in_bases
-        ),
-        "components": check.collect.with_unknown_components(
-            root, ns, brick_imports_in_components
-        ),
-    }
+from polylith import check, deps, info
 
 
 def calculate_diff(
@@ -29,7 +9,7 @@ def calculate_diff(
     project_data: dict,
     workspace_data: dict,
 ) -> dict:
-    brick_imports = get_brick_imports(root, namespace, project_data)
+    brick_imports = deps.get_brick_imports(root, namespace, project_data)
 
     all_bases = workspace_data["bases"]
     all_components = workspace_data["components"]
