+++ projects/poetry_polylith_plugin/pyproject.toml
@@ -15,6 +15,7 @@ packages = [
     {include = "polylith/check",from = "../../components"},
     {include = "polylith/commands",from = "../../components"},
     {include = "polylith/configuration",from = "../../components"},
+    {include = "polylith/deps",from = "../../components"},
     {include = "polylith/development",from = "../../components"},
     {include = "polylith/diff",from = "../../components"},
     {include = "polylith/dirs",from = "../../components"},
